refactor: extract classes
Move classes out into own files, promote them to production code.
Adjust some access modifier and scope of dependencies related to it
