+++ backup-stores/gcs/src/main/java/io/camunda/zeebe/backup/gcs/manifest/InvalidPersistedManifestState.java
